Home Work H12

+++ src/ru/progwards/java1/lessons/custing/AccuracyDoubleFloat.java
@@ -2,10 +2,8 @@ package ru.progwards.java1.lessons.custing;
 
 public class AccuracyDoubleFloat {
     public static double calculateAccuracy() {
-        double d = 1/3;
+        double d = ((double) 1 /3);
         float f = (float) d;
         return d - f;
     }
-
-
 }
